Break down module tests in two
