+++ tests/stubs/modules/authors/2019_10_18_193753_create_authors_tables.php
@@ -1,6 +1,5 @@
 <?php
 
-use Kalnoy\Nestedset\NestedSet;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 
@@ -58,29 +57,10 @@ class CreateAuthorsTables extends Migration
         Schema::create('author_revisions', function (Blueprint $table) {
             createDefaultRevisionsTableFields($table, 'author');
         });
-
-        Schema::create('author_categories', function (Blueprint $table) {
-            $table->string('name')->nullable();
-
-            // $table->nestedSet(); /// TODO /// This is not working inside tests
-
-            NestedSet::columns($table);
-        });
     }
 
     public function down()
     {
-        /**
-         * BUG in the nestedset package
-         *
-         * TODO
-         */
-        //        Schema::table('author_categories', function (Blueprint $table) {
-        //            $table->dropNestedSet();
-        //        });
-
-        Schema::dropIfExists('author_categories');
-
         Schema::dropIfExists('author_revisions');
 
         Schema::dropIfExists('author_translations');
