+++ tests/stubs/modules/authors/admin.php
@@ -3,3 +3,5 @@
 Route::group(['prefix' => 'personnel'], function () {
     Route::module('authors');
 });
+
+Route::module('categories');
