+++ tests/stubs/modules/authors/twill-navigation.php
@@ -11,4 +11,9 @@ return [
             ],
         ],
     ],
+
+    'categories' => [
+        'title' => 'Categories',
+        'route' => 'admin.categories.index',
+    ],
 ];
