+++ tests/stubs/modules/categories/CategoryRevision.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Models\Revisions;
+
+use A17\Twill\Models\Revision;
+
+class CategoryRevision extends Revision
+{
+    protected $table = 'category_revisions';
+}
