+++ tests/stubs/modules/categories/CategorySlug.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Models\Slugs;
+
+use A17\Twill\Models\Model;
+
+class CategorySlug extends Model
+{
+    protected $table = 'category_slugs';
+}
