+++ tests/stubs/modules/categories/form.blade.php
@@ -1,0 +1,10 @@
+@extends('twill::layouts.form')
+
+@section('contentFields')
+    @formField('input', [
+        'name' => 'description',
+        'label' => 'Description',
+        'translated' => true,
+        'maxlength' => 100
+    ])
+@stop
