+++ tests/stubs/modules/categories/site.category.blade.php
@@ -1,0 +1,1 @@
+Name: {{ $item->name }}
